enh(conflict_manager): fix troncated index_id in metrics table

+++ storage/inc/com/centreon/broker/storage/conflict_manager.hh
@@ -181,7 +181,7 @@ class conflict_manager {
   std::unordered_map<uint64_t, size_t> _cache_hst_cmd;
   std::unordered_map<std::pair<uint64_t, uint64_t>, size_t> _cache_svc_cmd;
   std::unordered_map<std::pair<uint64_t, uint64_t>, index_info> _index_cache;
-  std::unordered_map<std::pair<uint32_t, std::string>, metric_info>
+  std::unordered_map<std::pair<uint64_t, std::string>, metric_info>
       _metric_cache;
   std::mutex _metric_cache_m;
